Sun Sensivity now has better looking description
I am beyond words. It looked very ugly, both sentences being in the same line. Now that they are seperated, they are much nicer to look at. Flavor text is seperated from the actual effects of the trait in many other descriptions, so why not the same for sun sensivity? (Action hero, amputee, wildsman etc. to name a few)